Fixed leave-preview-mode-button link preload issue.
When using the next/link component and setting the href to a route handler, taht route handler will be called when nextjs preloads all internal links.

This caused multiple and unwanted calls to the draft-mode route handlers.

To prevent this, instead of the next/link a standard <a> tag should be used.